added erd
added erd and a readme just for database.

Thought a specialized readme might be helpful to people who might want to work on the database parts.

Great to explain the schema/ settings and what not.

Decided to separate this out because it might clutter up the root README since thats for more of a general overview